+++ App.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import {
+  SafeAreaView,
+  StatusBar,
+  useColorScheme,
+} from 'react-native';
+
+import Main from './screen/main/view-main';
+
+function App(): JSX.Element {
+  const isDarkMode = useColorScheme() === 'dark';
+  return (
+    <SafeAreaView>
+      <StatusBar
+        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
+        backgroundColor='white'
+      />
+      <Main />
+    </SafeAreaView>
+  );
+}
+
+export default App;
