+++ screen/main/view-main.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { View } from "react-native";
+import NavigationHeader from "./head-nav";
+import Post from "./post";
+import BottomNavigation from "./bottom-nav";
+
+ const Screen = () => {
+    return (
+        <View>
+            <NavigationHeader />
+            <Post />
+            <BottomNavigation/>
+        </View>
+    )
+ }
+
+export default  Screen
